Update configuration to point to new files
Now all configuration in ods-core points to the new file-structure
as given in the "configuration-sample" repository. Further, there
has been some cleanup of the templates.

Important is also the addition of a ODS_VERSION and ODS_GIT_REF param,
which allow to specify more easily what to tag and from which branch to
take the source code. For now, I use "latest" and "production" in order
not to break existing code but we need to think about how we handle that
in the future. For sure, on the release 2, we need to set at least
ODS_VERSION to 2, maybe even set ODS_GIT_REF to the branch "2.x" or so.