Add scripts to setup deployment target
create-target-project.sh:
This script can be used to create a new project, setup with serviceaccount
an image-puller secret so that the MRO can use it as a deploy target.
This script would typically be run by a cluster admin.

create-target-sa-secret.sh:
This script can be used to store the token of a serviceaccount with rights
to a deployment project in a secret, and share that secret with Jenkins.
This script would typically be run by a project admin.