Allow repo sync script to work with different target ref
For advanced users, who want to have a branch not named after one of the
official branches, this allows to use one branch as the source (e.g.
`3.x`) and merge/push that to a different target (e.g. `3.acme`).

It allows to have customisations in `3.acme`, while still updating with
the base branch (`3.x`).

This is intentionally not exposed in the `Makefile`.