Remove ODS config file
This wasn't used yet. It was intended to be used by the shared lib to
gain more information. However, in the future we might want to go to
fully prebuilt images (with UBI on the horizon), so we don't want to
lock ourselves in with adding this config in the image.