Use BITBUCKET_URL in Jenkins master
This works because the shared library can read both `BITBUCKET_URL` and
`BITBUCKET_HOST`.

The prov app will now send GIT_SERVER_URL as a full URL with scheme, so
we remove the detection if Git is HTTPS or not. Only a safeguard
defaulting to HTTPS stays to ease the transition time.