Add volumeMode field to PVC resources
This field has been added in Kubernetes 1.18, see
https://kubernetes.io/docs/concepts/storage/persistent-volumes/#volume-mode.

OpenShift 3.11 uses an older Kubernetes version, which ignores the field
if specified.

OpenShift 4 uses Kubernetes 1.18 and therefore defaults the field to
"Filesystem" if not specified, resulting in Tailor to see drift between
desired and current state.

To allow the template to be used against both OpenShift 3.11 and 4, we
now specify the field in the template, but preserve the live value.