Remove empty BuildConfig triggers
In OpenShift 3.11, they are automatically added to live configuration if
not specified in the template. Tailor ignores empty triggers during
diffing.

In OpenShift 4, empty triggers are removed in live configuration. During
diffing, Tailor detects drift as empty triggers are defined in the
template.

The easy solution here is to remove the empty BuildConfig triggers from
the templates.